requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
ipinfo==4.2.1
ipinfo-django==2.0.0
multidict==6.0.5
packaging==24.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
